local_macos/rag-sample/requirements.txt
flask==2.0.1
werkzeug==2.0.1
pymupdf==1.21.1
langchain==0.1.0
langchain-anthropic>=0.3.1
langchain-community>=0.0.10
anthropic>=0.41.1
chromadb==0.4.18
python-dotenv==0.19.2
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.36.2
